backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "market-pilot-backend"
version = "0.1.0"
description = "FastAPI backend for stock signal analysis, backtesting, and guarded execution."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "httpx>=0.27.0",
  "pydantic-settings>=2.2.1"
]

[tool.setuptools]
packages = ["app"]
